wip: api outline wip
